PathAttachmentModel updated
